+++ pytsite/widget/res/assets/js/datetime.js
@@ -2,7 +2,7 @@ $(function() {
     $('.widget-select-datetime').each(function() {
         var widget = $(this);
         widget.find('input').datetimepicker({
-            lang: pytsite.lang.current,
+            lang: pytsite.lang.current(),
             format: 'd.m.Y H:i',
             defaultDate: new Date()
         });